move each file to their separate folder
