+++ src/components/Button/Button.css
@@ -1,0 +1,4 @@
+.button{
+  padding: 10px;
+  width: 100px;
+}
