+++ src/components/ButtonPanel/ButtonPanel.css
@@ -1,0 +1,12 @@
+.panel{
+ display: grid;
+ grid-template-rows: repeat(4,1fr);
+ width: 700px;
+ align-items: center;
+
+}
+
+.panel .group{
+  display: flex;
+  justify-content: center;
+}
