+++ src/components/ButtonPanel/ButtonPanel.js
@@ -1,38 +1,41 @@
 import React from 'react';
-import Button from './Button';
+import Button from '../Button/Button';
+import '../ButtonPanel/ButtonPanel.css'
+
 
 const ButtonPanel = () => (
   <>
-    <div className="group1">
+  <div className='panel'>
+    <div className="group">
       <Button name="AC" />
       <Button name="+/-" />
       <Button name="%" />
       <Button name="÷" />
     </div>
-    <div className="group2">
+    <div className="group">
       <Button name="7" />
       <Button name="8" />
       <Button name="9" />
       <Button name="X" />
     </div>
-    <div className="group3">
+    <div className="group">
       <Button name="4" />
       <Button name="5" />
       <Button name="6" />
       <Button name="-" />
     </div>
-    <div className="group4">
+    <div className="group">
       <Button name="1" />
       <Button name="2" />
       <Button name="3" />
       <Button name="+" />
     </div>
-    <div className="group5">
+    <div className="group">
       <Button name="0" />
       <Button name="." />
       <Button name="=" />
     </div>
-
+  </div>
   </>
 );
